@@ -50,12 +50,12 @@
     "postcss-smart-import": "^0.7.6",
     "precss": "^4.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.5.2",
+    "react": "^16.9.0",
     "react-copy-to-clipboard": "^5.0.1",
     "react-datetime": "^2.16.3",
     "react-dnd": "^9.0.1",
     "react-dnd-html5-backend": "^9.0.0",
-    "react-dom": "^16.5.2",
+    "react-dom": "^16.9.0",
     "react-ga": "^2.4.1",
     "react-helmet": "^5.2.0",
     "react-intl": "^2.8.0",
