@@ -57,9 +57,9 @@
     "invariant": "^2.2.1",
     "lodash": "^4.17.5",
     "moment": "^2.16.0",
-    "react": "^16.0.0",
+    "react": "^16.9.0",
     "react-datetime": "^2.16.3",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.9.0",
     "react-intl": "^2.8.0",
     "react-router": "^5.0.0",
     "react-router-dom": "^5.0.0",
