@@ -18,8 +18,8 @@
     "invariant": "^2.2.1",
     "lodash": "^4.17.11",
     "pluralize": "^7.0.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-intl": "^2.8.0",
     "react-redux": "7.0.2",
     "react-router": "^5.0.0",
