@@ -20,9 +20,9 @@
     "lodash": "^4.17.11",
     "moment": "^2.24.0",
     "path-to-regexp": "^3.0.0",
-    "react": "^16.8.6",
+    "react": "^16.9.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.9.0",
     "react-intl": "^2.9.0",
     "react-redux": "^7.0.3",
     "react-router": "^5.0.0",
