@@ -14,9 +14,9 @@
     "immutable": "^3.8.2",
     "invariant": "^2.2.1",
     "lodash": "^4.17.11",
-    "react": "^16.0.0",
+    "react": "^16.9.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.0.0",
+    "react-dom": "^16.9.0",
     "react-intl": "^2.8.0",
     "react-redux": "^7.0.2",
     "react-router": "^5.0.0",
