@@ -21,8 +21,8 @@
     "koa2-ratelimit": "^0.9.0",
     "lodash": "^4.17.11",
     "purest": "^2.0.1",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-intl": "^2.8.0",
     "react-redux": "^7.0.2",
     "react-router": "^5.0.0",
