@@ -31,16 +31,22 @@ class ImgPreview extends React.Component {
   componentDidMount() {
     // We don't need the generateImgURL function here since the compo will
     // always have an init value here
-    const file = this.props.multiple ? get(this.props.files, ['0', 'name'], '') : get(this.props.files, 'name');
+    const file = this.props.multiple
+      ? get(this.props.files, ['0', 'name'], '')
+      : get(this.props.files, 'name');
     this.setState({
-      imgURL: get(this.props.files, ['0', 'url'], '') || get(this.props.files, 'url', ''),
+      imgURL:
+        get(this.props.files, ['0', 'url'], '') ||
+        get(this.props.files, 'url', ''),
       isImg: this.isPictureType(file),
     });
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (nextProps.isUploading !== this.props.isUploading) {
-      const lastFile = this.props.multiple ? nextProps.files.slice(-1)[0] : nextProps.files[0] || nextProps.files;
+      const lastFile = this.props.multiple
+        ? nextProps.files.slice(-1)[0]
+        : nextProps.files[0] || nextProps.files;
       this.generateImgURL(lastFile);
 
       if (this.props.multiple) {
@@ -48,7 +54,12 @@ class ImgPreview extends React.Component {
       }
     }
     // Update the preview or slide pictures or init the component
-    if (nextProps.didDeleteFile !== this.props.didDeleteFile || nextProps.position !== this.props.position || size(nextProps.files) !== size(this.props.files) && !this.state.isInitValue) {
+    if (
+      nextProps.didDeleteFile !== this.props.didDeleteFile ||
+      nextProps.position !== this.props.position ||
+      (size(nextProps.files) !== size(this.props.files) &&
+        !this.state.isInitValue)
+    ) {
       const file = nextProps.files[nextProps.position] || nextProps.files || '';
       this.generateImgURL(file);
 
@@ -66,14 +77,14 @@ class ImgPreview extends React.Component {
     console.log('An error occured in ImgPreview', info);
   }
 
-  getFileType = (fileName) => fileName.split('.').slice(-1)[0];
+  getFileType = fileName => fileName.split('.').slice(-1)[0];
 
   /**
    * [generateImgURL description]
    * @param  {FileList} files
    * @return {URL}
    */
-  generateImgURL = (file) => {
+  generateImgURL = file => {
     if (this.isPictureType(file.name) && !has(file, 'url')) {
       const reader = new FileReader();
       reader.onloadend = () => {
@@ -86,15 +97,16 @@ class ImgPreview extends React.Component {
       reader.readAsDataURL(file);
     } else if (has(file, 'url')) {
       const isImg = this.isPictureType(file.name);
-      const imgURL = file.url[0] === '/' ?`${strapi.backendURL}${file.url}` :file.url;
+      const imgURL =
+        file.url[0] === '/' ? `${strapi.backendURL}${file.url}` : file.url;
 
       this.setState({ isImg, imgURL });
     } else {
       this.setState({ isImg: false, imgURL: file.name });
     }
-  }
+  };
 
-  handleClick = (type) => {
+  handleClick = type => {
     const { files, position } = this.props;
     let file; // eslint-disable-line no-unused-vars
     let nextPosition;
@@ -109,51 +121,51 @@ class ImgPreview extends React.Component {
         nextPosition = files[position - 1] ? position - 1 : files.length - 1;
         break;
       default:
-        // Do nothing
+      // Do nothing
     }
 
     // Update the parent position
     this.updateFilePosition(nextPosition);
-  }
+  };
 
-  handleDragEnter = (e) => {
+  handleDragEnter = e => {
     e.preventDefault();
     e.stopPropagation();
     this.setState({ isDraging: true });
-  }
+  };
 
-  handleDragLeave = (e) => {
+  handleDragLeave = e => {
     e.preventDefault();
     e.stopPropagation();
     this.setState({ isDraging: false });
-  }
+  };
 
-  handleDragOver = (e) => {
+  handleDragOver = e => {
     e.preventDefault();
     e.stopPropagation();
-  }
+  };
 
-  handleDrop = (e) => {
+  handleDrop = e => {
     this.setState({ isDraging: false });
     this.props.onDrop(e);
-  }
+  };
 
   // TODO change logic to depend on the type
-  isPictureType = (fileName) => /\.(jpe?g|png|gif)$/i.test(fileName);
+  isPictureType = fileName => /\.(jpe?g|png|gif)$/i.test(fileName);
 
-  updateFilePosition = (newPosition) => {
+  updateFilePosition = newPosition => {
     this.props.updateFilePosition(newPosition);
-  }
+  };
 
   renderContent = () => {
     const fileType = this.getFileType(this.state.imgURL);
-    
-    if (this.state.isImg) {
-      const imgURL = startsWith(this.state.imgURL, '/') ? `${strapi.backendURL}${this.state.imgURL}` : this.state.imgURL;
 
-      return (
-        <img src={imgURL} alt="" />
-      );
+    if (this.state.isImg) {
+      const imgURL = startsWith(this.state.imgURL, '/')
+        ? `${strapi.backendURL}${this.state.imgURL}`
+        : this.state.imgURL;
+
+      return <img src={imgURL} alt="" />;
     }
 
     return (
@@ -161,25 +173,24 @@ class ImgPreview extends React.Component {
         <i className={`fa fa-file-${fileType}-o`} />
       </div>
     );
-  }
+  };
 
   render() {
     const { files, onBrowseClick } = this.props;
     const { imgURL } = this.state;
 
-    const containerStyle = isEmpty(imgURL) ?
-      {
-        backgroundImage: `url(${BkgImg})`,
-        backgroundRepeat: 'no-repeat',
-        backgroundPosition: 'center',
-        zIndex: 9999,
-      } : {};
+    const containerStyle = isEmpty(imgURL)
+      ? {
+          backgroundImage: `url(${BkgImg})`,
+          backgroundRepeat: 'no-repeat',
+          backgroundPosition: 'center',
+          zIndex: 9999,
+        }
+      : {};
 
     return (
       <div
-        className={cn(
-          styles.imgPreviewContainer,
-        )}
+        className={cn(styles.imgPreviewContainer)}
         onDragOver={this.handleDragOver}
         onDragEnter={this.handleDragEnter}
         style={containerStyle}
@@ -196,7 +207,7 @@ class ImgPreview extends React.Component {
           onDrop={this.handleDrop}
           showWhiteHint={this.state.isDraging || isEmpty(files)}
         />
-        { !isEmpty(imgURL) && this.renderContent() }
+        {!isEmpty(imgURL) && this.renderContent()}
         <ImgPreviewArrow
           enable={isArray(files) && size(files) > 1}
           onClick={this.handleClick}
@@ -230,10 +241,7 @@ ImgPreview.defaultProps = {
 
 ImgPreview.propTypes = {
   didDeleteFile: PropTypes.bool,
-  files: PropTypes.oneOfType([
-    PropTypes.object,
-    PropTypes.array,
-  ]),
+  files: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
   isUploading: PropTypes.bool,
   multiple: PropTypes.bool,
   onBrowseClick: PropTypes.func,
