@@ -21,7 +21,8 @@ import InputSpacer from '../InputSpacer';
 
 import styles from './styles.scss';
 
-class InputEmailWithErrors extends React.Component { // eslint-disable-line react/prefer-stateless-function
+class InputEmailWithErrors extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   state = { errors: [], hasInitialValue: false };
 
   componentDidMount() {
@@ -38,7 +39,7 @@ class InputEmailWithErrors extends React.Component { // eslint-disable-line reac
     }
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     // Show required error if the input's value is received after the compo is mounted
     if (!isEmpty(nextProps.value) && !this.state.hasInitialValue) {
       this.setState({ hasInitialValue: true });
@@ -59,10 +60,14 @@ class InputEmailWithErrors extends React.Component { // eslint-disable-line reac
   handleBlur = ({ target }) => {
     // Prevent from displaying error if the input is initially isEmpty
     if (!isEmpty(target.value) || this.state.hasInitialValue) {
-      const errors = validateInput(target.value, this.props.validations, 'email');
+      const errors = validateInput(
+        target.value,
+        this.props.validations,
+        'email'
+      );
       this.setState({ errors, hasInitialValue: true });
     }
-  }
+  };
 
   render() {
     const {
@@ -102,7 +107,7 @@ class InputEmailWithErrors extends React.Component { // eslint-disable-line reac
         className={cn(
           styles.containerEmail,
           this.props.customBootstrapClass,
-          !isEmpty(this.props.className) && this.props.className,
+          !isEmpty(this.props.className) && this.props.className
         )}
         style={style}
       >
@@ -134,7 +139,7 @@ class InputEmailWithErrors extends React.Component { // eslint-disable-line reac
         />
         <InputErrors
           className={errorsClassName}
-          errors={!noErrorsDescription && this.state.errors || []}
+          errors={(!noErrorsDescription && this.state.errors) || []}
           name={name}
           style={errorsStyle}
         />
@@ -205,10 +210,7 @@ InputEmailWithErrors.propTypes = {
   labelStyle: PropTypes.object,
   name: PropTypes.string.isRequired,
   noErrorsDescription: PropTypes.bool,
-  onBlur: PropTypes.oneOfType([
-    PropTypes.bool,
-    PropTypes.func,
-  ]),
+  onBlur: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
   onChange: PropTypes.func.isRequired,
   onFocus: PropTypes.func,
   placeholder: PropTypes.string,
