@@ -21,7 +21,8 @@ import InputSpacer from '../InputSpacer';
 
 import styles from './styles.scss';
 
-class InputSearchWithErrors extends React.Component { // eslint-disable-line react/prefer-stateless-function
+class InputSearchWithErrors extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   state = { errors: [], hasInitialValue: false };
 
   componentDidMount() {
@@ -38,7 +39,7 @@ class InputSearchWithErrors extends React.Component { // eslint-disable-line rea
     }
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     // Show required error if the input's value is received after the compo is mounted
     if (!isEmpty(nextProps.value) && !this.state.hasInitialValue) {
       this.setState({ hasInitialValue: true });
@@ -62,7 +63,7 @@ class InputSearchWithErrors extends React.Component { // eslint-disable-line rea
       const errors = validateInput(target.value, this.props.validations);
       this.setState({ errors, hasInitialValue: true });
     }
-  }
+  };
 
   render() {
     const {
@@ -102,7 +103,7 @@ class InputSearchWithErrors extends React.Component { // eslint-disable-line rea
         className={cn(
           styles.containerSearch,
           this.props.customBootstrapClass,
-          !isEmpty(this.props.className) && this.props.className,
+          !isEmpty(this.props.className) && this.props.className
         )}
         style={style}
       >
@@ -134,7 +135,7 @@ class InputSearchWithErrors extends React.Component { // eslint-disable-line rea
         />
         <InputErrors
           className={errorsClassName}
-          errors={!noErrorsDescription && this.state.errors || []}
+          errors={(!noErrorsDescription && this.state.errors) || []}
           name={name}
           style={errorsStyle}
         />
@@ -205,10 +206,7 @@ InputSearchWithErrors.propTypes = {
   labelStyle: PropTypes.object,
   name: PropTypes.string.isRequired,
   noErrorsDescription: PropTypes.bool,
-  onBlur: PropTypes.oneOfType([
-    PropTypes.bool,
-    PropTypes.func,
-  ]),
+  onBlur: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
   onChange: PropTypes.func.isRequired,
   onFocus: PropTypes.func,
   placeholder: PropTypes.string,
