@@ -1,8 +1,8 @@
 /**
-*
-* Plugin
-*
-*/
+ *
+ * Plugin
+ *
+ */
 
 import React from 'react';
 import PropTypes from 'prop-types';
@@ -14,19 +14,26 @@ import Controller from '../Controller';
 
 import styles from './styles.scss';
 
-class Plugin extends React.Component { // eslint-disable-line react/prefer-stateless-function
+class Plugin extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   state = { collapse: false };
 
   componentDidMount() {
     // Open the application's permissions section if there are APIs
-    if (this.props.name === 'application' && !isEmpty(get(this.props.plugin, 'controllers'))) {
+    if (
+      this.props.name === 'application' &&
+      !isEmpty(get(this.props.plugin, 'controllers'))
+    ) {
       this.props.changePluginSelected('application');
       this.setState({ collapse: !this.state.collapse });
     }
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.pluginSelected !== this.props.pluginSelected && nextProps.pluginSelected !== this.props.name) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    if (
+      nextProps.pluginSelected !== this.props.pluginSelected &&
+      nextProps.pluginSelected !== this.props.name
+    ) {
       this.context.resetShouldDisplayPoliciesHint();
       this.setState({ collapse: false });
     }
@@ -42,7 +49,7 @@ class Plugin extends React.Component { // eslint-disable-line react/prefer-state
     if (this.state.collapse) {
       this.context.resetShouldDisplayPoliciesHint();
     }
-  }
+  };
 
   render() {
     const divStyle = this.state.collapse ? { marginBottom: '.4rem' } : {};
@@ -59,16 +66,14 @@ class Plugin extends React.Component { // eslint-disable-line react/prefer-state
           <div>
             {this.props.name !== 'application' && (
               <div className={styles.iconContainer}>
-                {icon &&  <img src={icon} alt="icon" />}
+                {icon && <img src={icon} alt="icon" />}
               </div>
             )}
             <div className={styles.name}>{this.props.name}</div>
             &nbsp;—&nbsp;
             <div className={styles.description}>
               {this.props.name === 'application' ? (
-                <FormattedMessage
-                  id="users-permissions.Plugin.permissions.application.description"
-                />
+                <FormattedMessage id="users-permissions.Plugin.permissions.application.description" />
               ) : (
                 <FormattedMessage
                   id="users-permissions.Plugin.permissions.plugins.description"
@@ -77,21 +82,30 @@ class Plugin extends React.Component { // eslint-disable-line react/prefer-state
               )}
             </div>
           </div>
-          { emptyApplication && <div className={this.state.collapse ? styles.chevronUp : styles.chevronDown}></div> }
+          {emptyApplication && (
+            <div
+              className={
+                this.state.collapse ? styles.chevronUp : styles.chevronDown
+              }
+            ></div>
+          )}
         </div>
         <Collapse isOpen={this.state.collapse}>
           <div />
           <div className={styles.controllerContainer}>
-            {map(get(this.props.plugin, 'controllers'), (controllerActions, key) => (
-              <Controller
-                inputNamePath={`permissions.${this.props.name}`}
-                isOpen={this.state.collapse}
-                key={key}
-                name={key}
-                actions={controllerActions}
-                resetInputBackground={this.state.resetInputBackground}
-              />
-            ))}
+            {map(
+              get(this.props.plugin, 'controllers'),
+              (controllerActions, key) => (
+                <Controller
+                  inputNamePath={`permissions.${this.props.name}`}
+                  isOpen={this.state.collapse}
+                  key={key}
+                  name={key}
+                  actions={controllerActions}
+                  resetInputBackground={this.state.resetInputBackground}
+                />
+              )
+            )}
           </div>
         </Collapse>
       </div>
